add debug option for compile

+++ aitfw.py
@@ -5,12 +5,16 @@ from config import generator as gn
 import os
 import sys
 
-def compile():
-    #print('-------------------------')
-   # print(APP_CONFIG_HEADER)
-    #print(gn.generateConfig())
-    #print('-------------------------')
-    #print(gn.generateMakefile())
+COMPILE_OPTIONS = [
+    '',
+    'debug',
+]
+
+def compile(option):
+    if option not in COMPILE_OPTIONS:
+        os.system('echo \"error: not compile option \\\"{op}\\\"\"'.format(op=option))
+        return
+
     with open('./app.cfg', mode='w') as f:
         f.write(gn.generateConfig())
     with open('./Makefile.inc', mode='w') as f:
@@ -18,8 +22,9 @@ def compile():
     os.chdir('../')
     os.system('make app={app_name}'.format(app_name=APP_NAME))
     os.chdir('{dir_name}'.format(dir_name=APP_NAME))
-    os.remove('./app.cfg')
-    os.remove('./Makefile.inc')
+    if option != 'debug':
+        os.remove('./app.cfg')
+        os.remove('./Makefile.inc')
 
 def mkdir(dir_name):
     os.system('mkdir {dir_name}'.format(dir_name = dir_name))
@@ -27,16 +32,21 @@ def mkdir(dir_name):
     os.system('mkdir {dir_name}/include'.format(dir_name = dir_name))
 
 def usage():
-    output_str='Usage: aitfw [compile] [mkdir <dir_name>]'
+    output_str='Usage: aitfw [compile [option]] [mkdir <dir_name>]'
     os.system('echo \"{str}\"'.format(str=output_str))
 
 
 if __name__ == '__main__':
-    if len(sys.argv) < 2:
+    argc = len(sys.argv)
+    if argc < 2:
         usage()
     elif sys.argv[1] == 'compile':
-        compile()
-    elif sys.argv[1] == 'mkdir' and len(sys.argv) >= 3:
+        if argc >= 3:
+            option = sys.argv[2]
+        else:
+            option = ''
+        compile(option)
+    elif sys.argv[1] == 'mkdir' and argc >= 3:
         mkdir(sys.argv[2])
     else :
         usage()
